touch/focaltech: Restore feature settings as touch IC resets.
Add a recovery mechanism to restore all firmware settings when touch IC
resets unexpectedly. In each IRQ, the driver checks the byte2 value of
register 0xB2, it records all the firmware feature settings currently,
and the driver can compare them to settings of host side, and restores
the settings to firmware again if they are different.

Bug: 231678216
Test: Touch works well, please refer to the test method in
b/231678216#comment2

Signed-off-by: Mason Wang <masonwang@google.com>
Change-Id: Id5ca4f7f1dd246a9db0edc1b71d6fa23fb79af4b